updated a number of tests to use getRule() and removed the duplicated DI unit tests
